refactor: rename calcite hook file, fix usearcgiswidget hook casing (ALL-1589)
