+++ src/components/widgets/MapWidgets.tsx
@@ -1,7 +1,7 @@
 import Home from '@arcgis/core/widgets/Home';
 import BasemapToggle from '@arcgis/core/widgets/BasemapToggle';
 import Locate from '@arcgis/core/widgets/Locate';
-import useArcGISWidget from '../../hooks/UseArcGISWidget';
+import useArcGISWidget from '../../hooks/useArcGISWidget';
 
 // Arcgos JS API Widgets that are overlaid on the map
 const MapWidgets: React.FC = () => {
