+++ src/hooks/useArcGISWidget.tsx
